refactor: renames ApprovalService to PublicationManagerService for handling publication approval requests and events
