+++ Unicorn.Web/PublicationManagerService.Tests/PublicationManagerService.Tests.csproj
@@ -4,8 +4,8 @@
         <TargetFramework>net8.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
-        <AssemblyName>Unicorn.Web.ApprovalService.Tests</AssemblyName>
-        <RootNamespace>Unicorn.Web.ApprovalService.Tests</RootNamespace>
+        <AssemblyName>Unicorn.Web.PublicationManagerService.Tests</AssemblyName>
+        <RootNamespace>Unicorn.Web.PublicationManagerService.Tests</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
@@ -47,7 +47,7 @@
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\ApprovalService\ApprovalService.csproj" />
+      <ProjectReference Include="..\PublicationManagerService\PublicationManagerService.csproj" />
     </ItemGroup>
 
 </Project>
